--- src/main/java/teamkiim/koffeechat/member/Member.java
@@ -1,13 +1,0 @@
-package teamkiim.koffeechat.member;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-
-@Entity
-@Getter
-public class Member {
-
-    @Id @GeneratedValue
-    @Column(name="member_id")
-    private Long id;  // 회원 id
-}
