+++ src/main/java/teamkiim/koffeechat/post/dev/domain/DevPost.java
@@ -1,4 +1,4 @@
-package teamkiim.koffeechat.post.dev;
+package teamkiim.koffeechat.post.dev.domain;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
