+++ src/main/java/teamkiim/koffeechat/postlike/domain/repository/PostLikeRepository.java
@@ -1,8 +1,9 @@
-package teamkiim.koffeechat.postlike;
+package teamkiim.koffeechat.postlike.domain.repository;
 
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+import teamkiim.koffeechat.postlike.domain.PostLike;
 
 import java.util.List;
 import java.util.Optional;
